Move the Disclosure component out of the material app scope
Following the journey of our DisclosureControllable component, the
Disclosure component is moving a scope up with its sibling so that we
can use it between apps. It wouldn't make sense to have one disclosure
on the component level, and the other one in the material app scope.
